moved render trait to gosub_interface
