+++ crates/gosub_interface/src/lib.rs
@@ -7,4 +7,5 @@ pub mod html5;
 pub mod input;
 pub mod layout;
 pub mod node;
+pub mod render;
 pub mod request;
