+++ crates/gosub_render_pipeline/src/render/render_context.rs
@@ -1,11 +1,4 @@
-use crate::render::render_list::RenderList;
-use crate::render::viewport::Viewport;
+//! Re-export of the `RenderContext` trait, which now lives in
+//! `gosub_interface::render::render_context`.
 
-/// Abstraction over the per-tab state that render backends need.
-///
-/// Implemented by `gosub_engine::BrowsingContext`. Defined here so the render
-/// crate does not depend on `gosub_engine`.
-pub trait RenderContext {
-    fn viewport(&self) -> &Viewport;
-    fn render_list(&self) -> &RenderList;
-}
+pub use gosub_interface::render::render_context::*;
