+++ crates/gosub_renderer_skia/src/backend.rs
@@ -1,5 +1,5 @@
 use anyhow::{anyhow, Result};
-use gosub_render_pipeline::rasterizer::{RasterStrategy, Rasterable};
+use gosub_render_pipeline::rasterizer::{erase_rasterizer, RasterStrategy};
 use gosub_render_pipeline::render::backend::{
     ErasedSurface, ExternalHandle, PixelFormat, PresentMode, RenderBackend, RgbaImage, SurfaceSize,
 };
@@ -153,8 +153,8 @@ impl RenderBackend for SkiaBackend {
         })
     }
 
-    fn create_rasterizer(&self) -> Box<dyn Rasterable + Send + Sync> {
-        Box::new(crate::SkiaRasterizer::new(1.0))
+    fn create_rasterizer(&self) -> Box<dyn Any + Send + Sync> {
+        erase_rasterizer(Box::new(crate::SkiaRasterizer::new(1.0)))
     }
 
     fn raster_strategy(&self) -> RasterStrategy {
